telegram-course-bot/requirements.txt
fastapi
uvicorn
sqlalchemy
jinja2
python-multipart
python-dotenv
requests
telethon
googlesearch-python
beautifulsoup4

